Move schedule and map into features, remove unused images
