+++ src/features/Schedule/navigation.js
@@ -1,0 +1,18 @@
+import ScheduleContainer from './index';
+import EventDetails from './EventDetails';
+
+export default {
+  Schedule: {
+    screen: ScheduleContainer,
+    navigationOptions: {
+      title: 'SCHEDULE'
+    }
+  },
+  EventDetails: {
+    screen: EventDetails,
+    initialRouteParams: { },
+    navigationOptions: {
+      title: 'SCHEDULE',
+    }
+  }
+};
