+++ src/features/StaticMap/navigation.js
@@ -1,0 +1,11 @@
+import StaticMap from './index';
+
+export default {
+    StaticMap: {
+        screen: StaticMap,
+        initialRouteParams: { },
+        navigationOptions: {
+          title: 'MAP',
+        }
+      }
+};
